refactor: move method  from Receive to utils
